+++ cf/read_write/write.py
@@ -88,7 +88,7 @@ def write(fields, filename, fmt='NETCDF4', overwrite=True,
 
 
     **NetCDF hierarchical groups**
-    
+
     Hierarchical groups in CF provide a mechanism to structure
     variables within netCDF4 datasets with well defined rules for
     resolving references to out-of-group netCDF variables and